backend/requirements.txt
black==25.1.0
blinker==1.9.0
click==8.1.8
Flask==3.1.1
flask-cors==6.0.1
geographiclib==2.0
geopy==2.4.1
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
piexif==1.1.3
pillow==11.3.0
platformdirs==4.3.8
psycopg2-binary==2.9.9
python-dotenv==1.1.0
tomli==2.2.1
typing_extensions==4.14.0
Werkzeug==3.1.3
zipp==3.23.0
